plugins: Add Git objects graph import plugin
Add a new plugin to import the git objects (commits, trees, blobs) graph
from a local repository.

Branches and tags will also be imported as subgraphs of the commits graph.

Tree targeted by each commit can also be optionally extracted and added
as a subgraph.

The plugin implementation is based on the use of the great libgit2 library.