gha/talipot-python: Build and deploy universal macOS binary wheels
Add a new GitHub Actions workflow to build talipot-python binary wheels
for macOS with MACOSX_DEPLOYMENT_TARGET set to 10.9.

It enables to import the talipot module on every macOS >= 10.9, thus
supporting legacy versions of the OS.

The few required library dependencies will be compiled from source
(mostly using MacPorts) to ensure binary compatibilities.

Also deactivate the previous job on AppVeyor as the build timeout
of one hour is too low due to the amount of C/C++ code to compile.