plugins/TableView: Turn "Add property" menu entries to "Hide property"
In the context menus displayed when right-clicking on the horizontal header
of the table or in the Properties editor tab, turn the "Add new property"
entry into an "Hide property" one.

As the context menus are related to a selected property, it was weird to
have a non related action here.