talipot-python: Port talipot-core bindings to SIP >= 6.6
Port the talipot-core Python bindings from SIP 4.19 to SIP 6.6.
Nevertheless, keep compiling sip and talipot modules using cmake
instead of relying on SIP (as mingw is not supported).

From now on, SIP source code is no longer available in thirdparty,
the tool needs to be installed through pip or a package manager.

As the way SIP wraps enums in version 6 has changed (Python 3 enums
are now used), turn some talipot-core enums to C++11 scoped enums
in order to get the same namespaces in C++ and Python.