plugins/interactor: Add Graph Splatting interactor (WIP)
I stumbled across that old code of mine I thought was lost but the SVN
repository is in fact sill online and publicly available, below is the
checkout URL.
https://svn.labri.fr/repos/tulip_plugins/public/testing/interactors/GraphSplattingInteractor/

Let's try to make that fun Tulip plugin works again in Talipot.
After some port, it currently compile and does not crash but the visualization
it produces is no longer displayed. More debugging to do nut that's a good
start.