talipot-python: Rework exception management for talipot python plugins
When a talipot plugin written in Python is invoked from the classical Python
interpreter and not from the talipot GUI, prefer to let exception that might
be raised by its implementation continue its way instead of caughting it.

This behavior is much more pythonic than simply printing the error traceback
and clearing the python error indicator as client code calling the plugin can
catch the errors it might raise.