talipot-gui/QtGlSceneZoomAndPanAnimator: Speedup zoom and pan animation
Previously the graphChanged signal was emitted at each view refresh when
zooming and panning which was triggering a full redraw of the overview
widget when visible and thus was slowing down the zoom and pan animation.

As a full redraw of the overview widget is not needed when zomming and
panning, ensure the graphChanged signal is no longer emitted when
performing the animation.