plugins/TableView: Improve performance by removing useless processing
Visible rows and columns of the table are already resized to their content
each time the table is painted so there is no need to do that processing
each time the underlying model changes.

This greatly improve performance when a large number of graph elements have
their properties modified, using a Python script for instance, as each table
row were resized to their content previoulsy.