GeographicView: Simplify implementation for rendering online maps
Previous implementation was depending on QtWebKit or QtWebEngine to render
online maps using javascript and the open source Leaflet library.

While it was working fine, it was quite cumbersome to deploy as QtWebKit
is now deprecated and QtWebEngine is a beast in terms of size of binaries.

Prefer to use the lightweight QGeoView library instead which only depends
on Qt base libraries and works pretty fine to smoothly render an online map
and interact with it.

Inspired from the changes in the recent Tulip 6.0 release.